+++ plugin/src/main/java/com/denizenscript/denizen/scripts/commands/npc/EngageCommand.java
@@ -163,6 +163,6 @@ public class EngageCommand extends AbstractCommand {
      * @param duration the number of seconds to engage the NPCTag
      */
     public static void setEngaged(NPC npc, PlayerTag player, int duration) {
-        currentlyEngaged.put(getID(npc, player), CoreUtilities.monotonicMillis() + duration * 1000);
+        currentlyEngaged.put(getID(npc, player), CoreUtilities.monotonicMillis() + (long) duration * 1000L);
     }
 }
